Increase version of Gulp to 5 and other npm package wrangling.
There were a number of security issues, and this felt surprisingly complicated.
Firstly, node-sass was a dependency of Gulp 4 and required the version of
Node to be pinned to 14 in order to install cleanly.

Gulp 5 depends on (dart) sass or node-sass, but node-sass is deprecated:
https://www.npmjs.com/package/node-sass so we are moving to (Dart) 'sass'.
This probably removes the requirement to pin to version 14, but it's the LTS
so I don't see a reason to change this. In addition to this change, Gulp 5
will error out with an excellent error message to make the change shown in
gulpfile.js