Update test fixtures
This is the result of loading the test fixtures, migrating up, and re-exporting using the new export script. It applies the latest changes from how django handles exports, as well as the latest schema migrations.

This will reduce deltas in future whenever the actual content of test fixtures are updated.