Clean up multiple sources of nhs-frontend
Previously, we were loading the NHS Frontend CSS and JS twice; once was
coming from the wagtailnhsukfrontend package, and once was from a manual
copy of the nhs-frontend npm package which was kept in the `packages`
folder. This was a stale copy, with some custom CSS tweaks being
included, then being compiled and checked back into the repository for
`collectstatic` to find and prepare as part of the static assets
pipeline.

This change swaps to installing nhs-frontend from npm, and then
including that in the `nhsuk.scss` file in place of the previously
copied version. This means the compiled version of `nhsuk.css` is now
using the latest version (as pinned in package.json).

With the latest version available in the compiled CSS, we can drop the
lines in the `base.html` template which included the
wagtailnhsukfrontend version.