Configure code reporting to Codeclimate
This uses `coverage` to gather test coverage data for the project, then
reports to CodeClimate so we can get a better picture idea of where
we're not testing enough, and where new features have insufficient
tests.

This is intentionally a first pass, we'll probably want to start
excluding things and better configuring tests down the line.