Add custom main menu class to support captions
This overrides the default wagtailmenu item class with a custom one
which supports adding captions to menu items. The addition of captions
to the menu template is simple, but works in the existing structure (and
will be prettified when the main menu recieves its redesign).

This causes the existing menu structure to disappear, and it will need
to be re-imported. No changes are needed to the importer, since that
only cared about the model of the menu itself and not its item.