More precisely expose the search query in template
Previously this was relying on the `__str__` behaviour of a `Query` to
expose the string which was searched for to the template.

This is a poor representation of what we actually wanted `query` to do,
so this commit replaces it with a more explicit `query_string` ready for
promoted search results to be included in the template.