Nested loop no longer reuses variable name in loop body
In Python variables have function-wide scope which means that if two
variables have the same name in the same scope, they are in fact one
variable. Consequently, nested loops in which the target variables have
the same name in fact share a single variable. Such loops are difficult
to understand as the inner loop will modify the target variable of the
outer loop. This may lead to unexpected behavior if the loop variable is
used after the inner loop has terminated.