Change importer to be idempotent
Previously, the importer required a clean database, and would create new
objects for every piece of imported content.

This commit changes the behaviour so that the importer will now attempt
to match content found in the scrape to existing objects and perform an
in-place update. This means that it can be left to run on a schedule,
making transition into private beta and eventual migration smoother.