Enforce uniqueness of page/category relationships
There is no situation where a page and a category should be related more than once; enforce this constraint on the database level, and make `create_category_relationships_for_page` catch for (and silently discard) `IntegrityError`s. This is noticeably faster than `get_or_create`, which uses two queries for the operation, against just catching the error and ignoring it. The model is simple enough that we can't get integrity errors for anything else.