Correctly ignore cms/static directory and its contents
The contents of this file are compiled artefacts and assets brought from elsewhere in the site structure. This `collectstatic` process is run automatically at deploy-time, so there is no need for these to be checked in to the repository.

Modify the `.gitignore` to correctly ignore the `static` directory under `cms` instead of the root folder, and remove compiled assets from the codebase.