Add modified dates to all content pages
I have only tested the happy path for this at this moment. It doesn't
test if self.last_published_at == self.first_published_at. This is
mostly because I couldn't find a post which didn't have any
modifications made to it to visually inspect.
This commit only implements the logic for adding the modified dates.
The language used changes throughout the application - it is displayed
as `Lastest Version` and `Updated On`. I made the decision to use
Updated On, but a decision needs to be made about which one should be
used.
This also doesn't add in any styling for this logic either.
One for future commits.