Add one day to end of date range
The date range is actually a time range, and the time interpretation of a day is
the midnight that starts that day. So items posted on that day count for the lower bound
but not for the upper bound. This effectively moves the upper bound to 24:00 the same day.

This does raise questions about the naming of the fields: "published before/after" feel exclusive.