Improve asset pipeline
This makes some changes to how the asset pipeline works, namely that the `static` folder is now a place we can put our own static assets, and the output from `npm run build` will end up in `static_compiled`. Running `collectstatic` will then move these to `static_root` for serving or if running with `DEBUG` then they'll be served directly.

The upshot of this is we can include assets without having to mess about changing the asset compilation process itself.