Suppress 'naive datetime' warning
Whilst running the tests there was a lot of spam of the form:

RuntimeWarning: DateTimeField Page.latest_revision_created_at
received a naive datetime (9999-12-31 23:59:59.999999)
while time zone support is active.

This suppresses this warning.

There's concievably a bug where documents submitted around midnight
are in the wrong day due to BST, but that's a problem for later.