Search only for ticked publication types
We use Q objects to build up a long filter of "a OR b OR c" which we then
filter the existing queryset on. We needed to add searching on the
publication_type ids to suppress the following error:

Cannot filter search results with field "publication_type_id".
Please add index.FilterField('publication_type_id') to Publication.search_fields

We're using the slugs heavily in the front-end, rather than ids, mostly
because publication_type=5 isn't very human readable. There's currently
problems with the data appearing multiple times, but that's a data problem.

Additional documentation for the new publication_type fixtures are
in fixtures/publication_type.README.txt