Reformat document block for publication templates
We do this throughout the website -- this isn't just limited to
document collections/publications.

There's a new fixture containing links to external URLs, internal pages,
a 'pdf', 'docx' and random other file. (Really, the files are just
the string "x" because we're only checking the extension -- they're not
real so don't be surprised if you can't click through correctly.)

To do:
* Offsite link to image

To not do right now:
* remove Embed and Free Text components of the document group
* counting the pages in a PDF.