Keep example up to date with reality
We changed the number of replicas from 3 to 1 to see if it was causing issues.
It was; but seemed to require two `terraform apply`s to succeed since it didn't
have a valid SCRAPY_ENDPOINT after changing scrapy's settings, so didn't touch
the web.

We figured that the example file shouldn't have a broken state.