Repair migrations
Migrations were in a state which could not be applied cleanly. This removes the bad migrations as far back as the last known good migration on the test deployment, and recreates them.

If your database has had one of the bad migrations applied, you should:

```
./manage.py migrate publications 0003
./manage.py migrate pages 0007
```

to restore it to the known good state, and then `./manage.py migrate` forwards as normal.