Set Timezone to allow for DST
The time for when something is published is one hour out due to
British Summer Time. We set the timezone to Europe/London to
fix this.

We note that this does not appear to change the default time on the
automated publish/expiry selectors -- perhaps this is tied
to the user's browser's time?